Silence compiler warnings.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@17056 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/SVN/SVNLogHelper.cpp
@@ -24,7 +24,7 @@ BOOL SVNLogHelper::Log(LONG rev, const CString& author, const CString& /*date*/,
 	messages[rev] = message;
 	authors[rev] = author;
 	m_rev = rev;
-	for (int i=0; i<cpaths->GetCount(); ++i)
+	for (size_t i=0; i<cpaths->GetCount(); ++i)
 	{
 		const LogChangedPath& cpath = cpaths->GetAt(i);
 		if (m_relativeurl.Compare (cpath.GetPath())== 0)
